Allow cache timestamp format to be changed like in ActiveRecord.
