+++ spec/mongoid/document_spec.rb
@@ -122,75 +122,6 @@ describe Mongoid::Document do
     end
   end
 
-  describe "#cache_key" do
-
-    let(:document) do
-      Dokument.new
-    end
-
-    context "when the document is new" do
-
-      it "has a new key name" do
-        expect(document.cache_key).to eq("dokuments/new")
-      end
-    end
-
-    context "when persisted" do
-
-      before do
-        document.save
-      end
-
-      context "with updated_at" do
-
-        let!(:updated_at) do
-          document.updated_at.utc.to_s(:nsec)
-        end
-
-        it "has the id and updated_at key name" do
-          expect(document.cache_key).to eq("dokuments/#{document.id}-#{updated_at}")
-        end
-      end
-
-      context "without updated_at, with Timestamps" do
-
-        before do
-          document.updated_at = nil
-        end
-
-        it "has the id key name" do
-          expect(document.cache_key).to eq("dokuments/#{document.id}")
-        end
-      end
-    end
-
-    context "when model dont have Timestamps" do
-
-      let(:artist) do
-        Artist.create!
-      end
-
-      it "should have the id key name" do
-        expect(artist.cache_key).to eq("artists/#{artist.id}")
-      end
-    end
-
-    context "when model has Short Timestamps" do
-
-      let(:agent) do
-        ShortAgent.create!
-      end
-
-      let!(:updated_at) do
-        agent.updated_at.utc.to_s(:nsec)
-      end
-
-      it "has the id and updated_at key name" do
-        expect(agent.cache_key).to eq("short_agents/#{agent.id}-#{updated_at}")
-      end
-    end
-  end
-
   describe "#identity" do
 
     let(:person) do
